# Load view then make DB calls