--- a/template/transaction/partial/table.html
+++ b/template/transaction/partial/table.html
@@ -14,6 +14,12 @@
         {% for transaction in transactions %}
           {{ rows.transaction_row(transaction.tid, transaction.date, transaction.description, transaction.amount) }}
         {% endfor %}
+      {% else %}
+        <tr>
+          <td colspan="5">
+            <img src="static/img/three-dots.svg" width=100 />
+          </td>
+        </tr>
       {% endif %}
     </tbody>
   </thead>