… static/img/three-dots.svg